Making it looks a little better
