+++ client/src/components/Nav/index.js
@@ -1,128 +1,13 @@
-import React, { Component } from "react"
-import { Link, withRouter } from "react-router-dom"
-import styled from "styled-components"
+import React from "react"
+import { withRouter } from "react-router-dom"
 import { connect } from "react-redux"
 
 import { logout } from "../../redux/actions/auth"
 import { isProtectedPath } from "../../redux/helpers"
-import { boxShadowMixin } from "../../theme/mixins"
-
-import Dropdown from "./Dropdown"
-// import LandingPageNav from "../Pages/LandingPage/LandingPageNav"
+import { NavStyles } from "./styles"
+import UnauthenticatedLinks from "./UnauthenticatedLinks"
+import AuthenticatedLinks from "./AuthenticatedLinks"
 import GitHubSvg from "../icons/GitHubSvg"
-
-const NavStyles = styled.div`
-  background: ${props => props.theme.white};
-  min-height: ${props => props.theme.navHeight};
-
-  /* relative positioning so that nav-links-cta 
-     can position absolute with Nav as its parent */
-  position: relative;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  ${boxShadowMixin};
-
-  padding-left: 2rem;
-  padding-right: 1.25rem;
-
-  .nav-links-wrapper {
-    /* This is to offset the Github logo bc of its absolute position */
-    margin-right: 60px;
-  }
-  .logo {
-    color: ${props => props.theme.primary};
-    font-weight: 700;
-    font-size: 1.75rem;
-    letter-spacing: -0.0275rem;
-  }
-
-  ul {
-    display: flex;
-    flex-direction: row;
-  }
-
-  li {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-  }
-
-  a {
-    display: flex;
-    align-items: center;
-    height: 3.125rem;
-    padding-left: 8px;
-    padding-right: 8px;
-    color: ${props => props.theme.midGray};
-    font-size: 1.0625rem;
-    font-weight: 400;
-    &:hover {
-      color: ${props => props.theme.primary};
-    }
-  }
-
-  button {
-    cursor: pointer;
-    border: none;
-    &:hover {
-      text-decoration: underline;
-    }
-    padding-top: 0;
-    padding-bottom: 0;
-  }
-
-  ul.authenticated-links {
-    li {
-      padding-left: 14px;
-      padding-right: 14px;
-    }
-  }
-
-  .call-to-action {
-    position: absolute;
-    top: 0;
-    right: 1.25rem;
-    height: 100%;
-    & a {
-      height: 100%;
-    }
-  }
-`
-
-const UnauthenticatedLinks = ({ pathname }) => (
-  <ul className="unauthenticated-links">
-    {pathname === "/login" || pathname === "/" ? (
-      <Link to="/signup">Sign up</Link>
-    ) : null}
-    {pathname === "/signup" || pathname === "/" ? (
-      <Link to="/login">Login</Link>
-    ) : null}
-  </ul>
-)
-
-class AuthenticatedLinks extends Component {
-  state = {}
-
-  handleLogout = e => {
-    e.preventDefault()
-    this.props.logout()
-  }
-
-  render() {
-    return (
-      <ul className="authenticated-links">
-        <Dropdown hidden={false} />
-        <li>
-          <a href="/logout" onClick={this.handleLogout}>
-            Log out
-          </a>
-        </li>
-      </ul>
-    )
-  }
-}
 
 const Nav = ({ location, logout, isLoggedIn }) => {
   const { pathname } = location
@@ -131,9 +16,6 @@ const Nav = ({ location, logout, isLoggedIn }) => {
 
   return (
     <div>
-      {/* {pathname === "/" ? (
-        <LandingPageNav />
-      ) : ( */}
       <NavStyles>
         <div className="logo">Backwoods Tracker</div>
         <div className="nav-links-wrapper">
@@ -146,7 +28,6 @@ const Nav = ({ location, logout, isLoggedIn }) => {
           <GitHubSvg width="32px" height="32px" />
         </div>
       </NavStyles>
-      {/* )} */}
     </div>
   )
 }
